Refactor and clean up in preparation for linen 2.0
- rename communities service to channel
- Extract buildthreads to buildfeed
