+++ nextjs/pages/subdomain/[communityName]/c/[channelName]/index.tsx
@@ -1,6 +1,6 @@
 import { GetStaticPropsContext } from 'next/types';
 import Channel from '../../../../../components/Pages/Channels';
-import { channelGetStaticProps } from '../../../../../services/communities';
+import { channelGetStaticProps } from '../../../../../services/channel';
 
 export async function getStaticProps(context: GetStaticPropsContext) {
   return channelGetStaticProps(context, true);
